時間計測処理の場所と内容を修正、time変数の名前を変更、不要なコードを削除
